Druid Feral - updated some spell values. Refactored AdaptiveSwarm modules into a single file now that it's Feral only
