+++ src/analysis/retail/druid/feral/modules/spells/LionsStrength.tsx
@@ -17,8 +17,6 @@ const LIONS_STRENGTH_BOOSTED_SPELLS = [
   SPELLS.FEROCIOUS_BITE,
   SPELLS.RAVAGE_DOTC_CAT,
   SPELLS.TEAR,
-  SPELLS.RAMPANT_FEROCITY,
-  // TODO TWW - Dreadful Wound? (probably not)
 ];
 
 /**
